Add guides section with the How to Build on Ethereum hub
